feat(abstract-utxo): extract signer keychain earlier
Do the whole decode dance before calling signTransaction

Issue: BTC-1450
